pundit done

+++ app/controllers/order_items_controller.rb
@@ -10,12 +10,15 @@ class OrderItemsController < ApplicationController
     order_item.order = order
     order_item.meal = meal
     order_item.unit_price = meal.offer_price
+    authorize order_item
+
     order_item.save
     redirect_to orders_path
   end
 
   def destroy
     order_item = OrderItem.find(params[:id])
+    authorize order_item
     order_item.destroy
     redirect_to meals_path
   end
